Rozróżnianie rodzajów zawodów w controllerze

+++ src/main/java/com/stelmyit/skijumping/competition/calculator/service/distance/DistanceHillValueService.java
@@ -7,34 +7,34 @@ import java.math.BigDecimal;
 @Service
 public class DistanceHillValueService {
 
-    public BigDecimal getMeterValue(final int kPoint) {
+    public float getMeterValue(final int kPoint) {
         if (kPoint < 25) {
-            return BigDecimal.valueOf(4.8);
+            return 4.8f;
         } else if (kPoint < 30) {
-            return BigDecimal.valueOf(4.4);
+            return 4.4f;
         } else if (kPoint < 35) {
-            return BigDecimal.valueOf(4);
+            return 4f;
         } else if (kPoint < 40) {
-            return BigDecimal.valueOf(3.6);
+            return 3.6f;
         } else if (kPoint < 50) {
-            return BigDecimal.valueOf(3.2);
+            return 3.2f;
         } else if (kPoint < 60) {
-            return BigDecimal.valueOf(2.8);
+            return 2.8f;
         } else if (kPoint < 70) {
-            return BigDecimal.valueOf(2.4);
+            return 2.4f;
         } else if (kPoint < 80) {
-            return BigDecimal.valueOf(2.2);
+            return 2.2f;
         } else if (kPoint < 100) {
-            return BigDecimal.valueOf(2);
+            return 2f;
         } else if (kPoint < 170) {
-            return BigDecimal.valueOf(1.8);
+            return 1.8f;
         } else  {
-            return BigDecimal.valueOf(1.2);
+            return 1.2f;
         }
     }
 
-    public BigDecimal getBaseDistancePoints(final int kPoint) {
-        return kPoint > 170 ? BigDecimal.valueOf(120) : BigDecimal.valueOf(60);
+    public int getBaseDistancePoints(final int kPoint) {
+        return kPoint > 170 ? 120 : 60;
     }
 }
