+++ src/test/java/com/stelmyit/skijumping/competition/calculator/service/distance/DistanceHillValueServiceTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.math.BigDecimal;
 import java.util.stream.Stream;
 
 import static java.util.stream.Stream.of;
@@ -18,9 +17,9 @@ public class DistanceHillValueServiceTest {
 
     @ParameterizedTest
     @MethodSource("meterValueTestData")
-    public void shouldCalculateMeterValue(final int kPoint, final BigDecimal expectedValue) {
+    public void shouldCalculateMeterValue(final int kPoint, final float expectedValue) {
         // when
-        final BigDecimal result = uut.getMeterValue(kPoint);
+        final float result = uut.getMeterValue(kPoint);
 
         // then
         assertThat(result).isEqualTo(expectedValue);
@@ -32,10 +31,10 @@ public class DistanceHillValueServiceTest {
         final int kPoint = 185;
 
         // when
-        final BigDecimal result = uut.getBaseDistancePoints(kPoint);
+        final int result = uut.getBaseDistancePoints(kPoint);
 
         // then
-        assertThat(result).isEqualTo(BigDecimal.valueOf(120));
+        assertThat(result).isEqualTo(120);
     }
 
     @Test
@@ -44,34 +43,34 @@ public class DistanceHillValueServiceTest {
         final int kPoint = 125;
 
         // when
-        final BigDecimal result = uut.getBaseDistancePoints(kPoint);
+        final int result = uut.getBaseDistancePoints(kPoint);
 
         // then
-        assertThat(result).isEqualTo(BigDecimal.valueOf(60));
+        assertThat(result).isEqualTo(60);
     }
 
     private static Stream<Arguments> meterValueTestData() {
         return of(
-            arguments(24, BigDecimal.valueOf(4.8)),
-            arguments(25, BigDecimal.valueOf(4.4)),
-            arguments(29, BigDecimal.valueOf(4.4)),
-            arguments(30, BigDecimal.valueOf(4)),
-            arguments(34, BigDecimal.valueOf(4)),
-            arguments(35, BigDecimal.valueOf(3.6)),
-            arguments(39, BigDecimal.valueOf(3.6)),
-            arguments(40, BigDecimal.valueOf(3.2)),
-            arguments(49, BigDecimal.valueOf(3.2)),
-            arguments(50, BigDecimal.valueOf(2.8)),
-            arguments(59, BigDecimal.valueOf(2.8)),
-            arguments(60, BigDecimal.valueOf(2.4)),
-            arguments(69, BigDecimal.valueOf(2.4)),
-            arguments(70, BigDecimal.valueOf(2.2)),
-            arguments(79, BigDecimal.valueOf(2.2)),
-            arguments(80, BigDecimal.valueOf(2)),
-            arguments(99, BigDecimal.valueOf(2)),
-            arguments(100, BigDecimal.valueOf(1.8)),
-            arguments(169, BigDecimal.valueOf(1.8)),
-            arguments(170, BigDecimal.valueOf(1.2))
+            arguments(24, 4.8f),
+            arguments(25, 4.4f),
+            arguments(29, 4.4f),
+            arguments(30, 4f),
+            arguments(34, 4f),
+            arguments(35, 3.6f),
+            arguments(39, 3.6f),
+            arguments(40, 3.2f),
+            arguments(49, 3.2f),
+            arguments(50, 2.8f),
+            arguments(59, 2.8f),
+            arguments(60, 2.4f),
+            arguments(69, 2.4f),
+            arguments(70, 2.2f),
+            arguments(79, 2.2f),
+            arguments(80, 2f),
+            arguments(99, 2f),
+            arguments(100, 1.8f),
+            arguments(169, 1.8f),
+            arguments(170, 1.2f)
         );
     }
